Do not overrun plugin string when copying it.
parse_plugin() expands the % in plugin string to plugin_copy.  It
checks that is does not write behind the end of the destination
memory, but the source length was not checked.  This resulted in
reading from a possibly invalid memory location which may cause a
segmentation fault.  Add a check for the string length of the source.